Round-6 red-team fixes: EF decimal materialization, crashed-forward liveness
Model-diverse round-6 sweep (Opus + Sonnet) on 1a15815.

- HIGH (Opus, empirically confirmed against LocalDB + EF6.2.0): the issuance tick's
  IntentRow read FeeInPercentage as double from a decimal(9,4) column. EF6 SqlQuery<T>
  does not widen decimal->double — it throws on materialization. LoadByStatus(Issuing)
  is the only Issuing loader, so it threw every tick, was swallowed by the outer catch,
  and NOTHING issued — while the pump kept moving USDT into the one-way float. Invisible
  to build + unit tests (runtime EF materialization). Fixed: IntentRow.FeeInPercentage is
  now decimal (matching the column), narrowed to double at the order assignment.
- MEDIUM (Sonnet): a createTransfer() throw left the refill row 'Initiated' (UpdateStatus
  never ran); the pump didn't handle 'Initiated', so the intent stuck in Funding forever,
  occupying the user's one-open slot with no self-service recovery. Root-caused:
  AttemptTransfer now wraps createTransfer in try/catch and records Unknown (ambiguous,
  may have moved) + alerts; the pump also treats Initiated like Unknown (advance to
  Confirming, where the evidence gate parks it for the 180-min reconcile alert).
- LOW (Sonnet): guard cardholder creation on holderId > 0 (a pass_audit with holderId 0
  would persist a bogus holder that fails at openCard).
- LOW (Sonnet): status-string drift — added a pinning unit test that fails the build if any
  CardFundingStatuses value or gate key changes, forcing the coupled raw-SQL literals and
  migration 0011 to be updated in lockstep (chosen over concatenating constants into ~8 SQL
  statements, since the migration hardcodes the same DB vocabulary anyway).

Accepted/documented (Opus LOW, not a blocker): forward-before-debit can strand a card's
float draw if the wallet is drained between Funding and Issuing via the direct-spend path;
already alerted + surfaced by the stuck sweep, and the direct-spend path is removed under
the streaming model (rollout).

Verified: QryptoCard.Sec/INT/INT.Callback build clean; 180 unit tests pass.

+++ QryptoCard.INT/Script/Service/CardFundingIssuanceService.cs
@@ -68,7 +68,7 @@ namespace QryptoCard.INT.Script.Service
                 HolderID = it.HolderID,
                 InitialDeposit = (double)it.Face,
                 Price = (double)it.Price,
-                FeeInPercentage = it.FeeInPercentage,
+                FeeInPercentage = (double)it.FeeInPercentage,
                 Fee = (double)it.PercentageFee,
                 // Debit the FULL amount the customer deposited for this card (price + face + % fee +
                 // fixed fee). Only `face` funds the card at WasabiCard; the rest is our margin/cost,
@@ -119,7 +119,7 @@ namespace QryptoCard.INT.Script.Service
                 UserID = it.UserID,
                 CardNo = it.CardNo,
                 Amount = (double)it.Face,
-                FeeInPercentage = it.FeeInPercentage,
+                FeeInPercentage = (double)it.FeeInPercentage,
                 Fee = (double)it.PercentageFee,
                 Total = (double)it.ExpectedTotal, // see IssueNew: full deposit, so no residual gifting
                 ReceivedAmount = (double)it.Face,
@@ -281,7 +281,10 @@ namespace QryptoCard.INT.Script.Service
             public string CardNo { get; set; }
             public decimal Face { get; set; }
             public decimal Price { get; set; }
-            public double FeeInPercentage { get; set; }
+            // MUST be decimal to match the FeeInPercentage decimal(9,4) column: EF6 SqlQuery<T> does NOT
+            // widen decimal->double, it throws on materialization. Reading it as double silently killed
+            // the whole issuance tick every run (RT round 6). Narrowed to double at the assignment.
+            public decimal FeeInPercentage { get; set; }
             public decimal PercentageFee { get; set; }
             public decimal ExpectedTotal { get; set; }
         }
